Start group editor refactor
Most things are working, but not everything's been checked.

Could probably still move tmpGroup to the editor component with events
for load-editor, editor-loaded etc.. Also a bunch of other stuff that
could probably be moved over to the component, we'll see when it works.
